requirements.txt
Flask==1.0.3
envparse==0.2.0
requests==2.21.0
Flask-Cors==3.0.7
flask-mongoengine==0.9.5
flask-redis==0.4.0
PyJWT==1.7.1
celery==4.3.0
pytest==5.0.0
marshmallow==3.0.0rc4
timber==2.1.0
gunicorn==19.9.0
blinker==1.4